Wrap high merch in standard container
