+++ src/core/create-ad-slot.ts
@@ -127,6 +127,24 @@ const concatSizeMappings = (
 		{ ...defaultSizeMappings },
 	);
 
+type ContainerOptions = {
+	className?: string;
+};
+
+const adSlotContainerClass = 'ad-slot-container';
+
+const wrapSlotInContainer = (
+	adSlot: HTMLElement,
+	options: ContainerOptions = {},
+) => {
+	const container = document.createElement('div');
+
+	container.className = `${adSlotContainerClass} ${options.className ?? ''}`;
+
+	container.appendChild(adSlot);
+	return container;
+};
+
 const createAdSlot = (
 	name: SlotName,
 	options: CreateSlotOptions = {},
@@ -150,4 +168,10 @@ const createAdSlot = (
 	);
 };
 
-export { createAdSlot, concatSizeMappings };
+export {
+	adSlotContainerClass,
+	concatSizeMappings,
+	createAdSlot,
+	wrapSlotInContainer,
+};
+export type { ContainerOptions };
